requirements.txt
# ============================================================
# Llais Cymraeg — Requirements
# Welsh Voice AI Stack
# Install with: pip install -r requirements.txt
# ============================================================

# --- Core Utilities ---
python-dotenv>=1.0.0       # Load secrets from .env files
tqdm>=4.66.0               # Progress bars for long data operations
numpy>=1.26.0              # Numerical computing (arrays, audio math)
requests>=2.31.0           # HTTP requests to APIs

# --- Audio Processing ---
soundfile>=0.12.1          # Read and write audio files (WAV, FLAC, etc.)
librosa>=0.10.1            # Audio analysis — tempo, pitch, spectrograms
pedalboard>=0.9.0          # Spotify's audio effects library (EQ, reverb, compress)

# --- Data Sources ---
datacollective>=0.1.0      # Official Mozilla Data Collective SDK
huggingface_hub>=0.23.0    # Upload/download models to HF Hub
datasets>=2.19.0           # Load datasets locally
transformers>=4.41.0       # Base transformer models (Whisper, etc.)

# --- Speech-to-Text (Phase 2) ---
faster-whisper>=1.0.0      # Faster, optimised Whisper runtime
silero-vad>=4.0.0          # Voice Activity Detection (detects when someone speaks)
jiwer>=3.0.0               # Word Error Rate (WER) evaluation metric

# --- Text-to-Speech (Phase 3) ---
TTS>=0.22.0                # Coqui XTTS v2 — Welsh voice cloning

# --- Language Model (Phase 4) ---
langchain>=0.2.0           # Orchestration framework for LLM pipelines
langchain-community>=0.2.0 # Community integrations (Ollama, ChromaDB)
chromadb>=0.5.0            # Vector database for conversation memory
ollama>=0.2.0              # Local LLM runtime (runs Llama 3.1 8B)

# --- API Layer (Phase 5) ---
fastapi>=0.111.0           # Modern Python web framework for our REST API
uvicorn>=0.29.0            # ASGI server that runs FastAPI
websockets>=12.0           # WebSocket support for real-time audio streaming
redis>=5.0.0               # Redis client for request queue management
python-multipart>=0.0.9    # Required by FastAPI for audio file uploads
